Conflicts:
	.gitignore
	android/.idea/findbugs-idea.xml
	android/res/drawable-nodpi/riviera.jpg
	android/res/raw/riviera_mtl
	android/res/raw/riviera_obj
	android/src/net/gtamps/android/core/renderer/graph/RenderableNode.java
	android/src/net/gtamps/android/core/renderer/graph/primitives/Camera.java
	android/src/net/gtamps/android/core/renderer/graph/primitives/Light.java
	android/src/net/gtamps/android/game/Game.java
	android/src/net/gtamps/android/game/objects/Object3dFactory.java
	android/src/net/gtamps/android/game/scene/Hud.java
	android/src/net/gtamps/android/game/scene/Scene.java
	android/src/net/gtamps/android/game/scenes/World.java
	shared/src/net/gtamps/shared/Config.java

+++ android/src/net/gtamps/android/core/renderer/mesh/buffermanager/VertexManager.java
@@ -66,7 +66,7 @@ public class VertexManager {
     public short addVertex(
             float pointX, float pointY, float pointZ,
             float normalX, float normalY, float normalZ,
-            float colorR, float colorG, float colorB, float colorA,
+            short colorR, short colorG, short colorB, short colorA,
             float textureU, float textureV) {
 
         vertices.add(pointX, pointY, pointZ);
